tools/job_manager/requirements.txt
python-jenkins==1.7.0
PyYAML==6.0